+++ CHANGELOG.md
@@ -1,1 +1,6 @@
 # Changelog
+
+## [1.0.1] - 2021-05-20
+### Fixed
+- Issue with parsing killstreak Unusual weapon names on backpack.tf.
+- Issue with links not auto-saving.